isolate each publishing
